Committing out of sheer amazement at how poorly documented the modules I settled on are about how to manage pixel data in arrays. Wild stuff.
What I found is that in order to correctly turn a Uint8ClampedArray of 4 channel pixel data into an ndarray, one needs to cast to the array with the following pattern.

```var myndarray = ndarray(myuint8clampedarray, [imgWidth, imgHeight, 4], [4, imgWidth*4, 1])```

My confusion lay in specifically what the "stride" of the array needed to be 4 then 4x the width of the image and then 1. The documentation did not explain what these shape definitions needed to correspond to for common usage.